win32: fix missing and misordered headers
Most headers in the net/ subdirectory were incorrect, and some defines
(like the utsname one) did not exist.

Also, in mesh.cpp, we did not include the correct header for
std::back_inserter(), which was <iterator>

There were some incorrect things happening with errno: we defined a
macro that pointed to a function call, but we did this in a header file
and this was causing some issues with other headers. We moved this
define to the cpp files to restrict them only when they were needed